Move functions to folders.
